+++ api/indexes.py
@@ -1,31 +1,11 @@
-EventPageMapping = {
+GenericMapping = {
     'properties': {
         'id': {'type': 'keyword'},
+        'type': {'type': 'keyword'},
         'name': {'type': 'text'},
-        'dtype': {'type': 'text'},
-        'location': {'type': 'text'},
-        'summary': {'type': 'text'},
+        'body': { 'type': 'text'},
         'date': {'type': 'date'},
     }
 }
 
-AppealPageMapping = {
-    'properties': {
-        'id': {'type': 'keyword'},
-        'name': {'type': 'text'},
-        'dtype': {'type': 'text'},
-        'location': {'type': 'text'},
-        'date': {'type': 'date'},
-    }
-}
-
-ReportPageMapping = {
-    'properties': {
-        'id': {'type': 'text'},
-        'name': {'type': 'text'},
-        'dtype': {'type': 'text'},
-        'location': {'type': 'text'},
-        'summary': {'type': 'text'},
-        'date': {'type': 'date'},
-    }
-}
+ES_PAGE_NAME = 'page_all'